requirements.txt
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
Django==3.2
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.5.1
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
wrapt==1.13.3
